Implement code coverage functionality
To enable code coverage, pass `-DUSE_COVERAGE=ON` to cmake. This will
enable the code to be compiled with the `-coverage` option.

In order to actually collect coverage information, build the project,
run the tests (either via ctest or directly), and then run
`./lcov-capture.sh`. This should result in a `coverage` directory which
will contain an `index.html` page to be viewed.

The script is generated for each build directory.